refactor : ElementCollection -> Converter 변경
엔티티 컬럼타입 변경 Long->Integer
- Integer 범위로도 충분할 것 같아 변경함.

AuthService.register ->
name과 moblieNumber/businessNumber 중복체크 로직을 하나로 통일
- existsByNameOrMobileNumber
- existsByNameOrBusinessNumber
- 쿼리 한번 절약

+++ src/main/java/com/devhong/free_coupon/dto/TemplateDto.java
@@ -21,9 +21,9 @@ public class TemplateDto {
         @NotBlank
         private String couponName;
         @NotNull
-        private Long couponValue;
+        private Integer couponValue;
         @NotNull
-        private Long expiredPeriod;
+        private Integer expiredPeriod;
         @NotBlank
         private String description;
         @NotBlank
@@ -62,8 +62,8 @@ public class TemplateDto {
         private Long template_id;
         private String category;
         private String couponName;
-        private Long couponValue;
-        private Long expiredPeriod;
+        private Integer couponValue;
+        private Integer expiredPeriod;
         private String description;
         private String imgUrl;
